+++ Project/include/sphere.hpp
@@ -9,10 +9,10 @@
 
 struct Sphere : public Hitable
 {
-    Vec3d center;
-    double radius = 1;
+    Vec3f center;
+    float radius = 1;
 
-    Sphere(Vec3d center, double radius, const Material* mat = nullptr): Hitable{mat},
+    Sphere(Vec3f center, float radius, const Material* mat = nullptr): Hitable{mat},
         center{center}, radius{radius} {}
 
     /**
